chore: remove npx from wc-dev (#12342)
Remove the usage of `npx`. When the `ui5nps` bin is not found, `npx` attempts to fetch it from the npm registry, which leads to an error. If running the scripts still results in a missing `ui5nps` bin, the dependencies need to be reinstalled. This can be done by removing `node_modules` and installing everything again, since Yarn does not automatically re-link new binaries unless the installation is refreshed.