feat(ui5-select): visible second column value and tooltip (#12390)
When read-only Select is displayed in a two column layout now the value of the Select includes the "text" and the "additionalText" of the selected item, separated by a character, provided by SelectTwoColumnSeparator enumeration. Possible values are Dash, Bullet and VerticalLine.

When Select is read-only there is a default tooltip shown if text is truncated. The default tooltip is the "text" of the item + "additionalText", if any, separated by "-".

If the user has already set "tooltip", it has precedence over the default tooltip. In this case, the default tooltip is not provided.