fix(ui5-dynamic-date-range): focus last selected option, upon navigate back (#12233)
Previously in our ui5-dynamic-date-range component when user selects an option with Template (e.g. SingleDate, DateRange, Last/Next, etc.) from the List and then clicking the Back Button, when the view switches back to the List with the options, the last clicked option is not being focused.

Now clicking the Back Button no matter if the user selected the date/time or not, switching the view to the List view focus the last clicked list item.

Why we used this approach:
When navigating back, the component re-renders to show the options list.
The focus attempt must happen after the list items are fully rendered in the DOM
Without the await renderFinished and setTimout() in the onAfterRendering hook, this._list.items are empty or the DOM elements might not be ready for focus.