fix(ui5-datetime-picker): restore last valid value (#12202)
Previously our in ui5-datetime-picker when a user enters an invalid value (e.g., "invalid text") and then selects a new date from the calendar, the time portion was being reset to 12:00:00 AM instead of preserving the previously valid time.

The reason behind that is that the component was not storing the last valid time value.
So, when a new date was selected, the component had no reference to the previously valid time, so it defaulted to midnight (12:00:00 AM).

We now store the valid value and fallback to it whenever needed. And if no value was provided, the default behaviour is preserved (auto selects the current time)