feat(ui5-table): alternateRowColors property added (#12468)
* feat(ui5-table): alternateRowColors property added

In current design it is available only for regular state of the rows.
Fixes: #11597

Accessibility improvements for the growing button.
Although announcement is OK a button should not placed in a grid.
Fixes: #11143

* feat(ui5-table): alternateRowColors property added

In current design it is available only for regular state of the rows.
Fixes: #11597

Accessibility improvements for the growing button.
Growing button is now externalized from the grid.
Fixes: #11143