fix(ui5-calendar): remove auto-focus logic from onAfterRendering() hook (#12465)
With this change we remove the focusing logic in the onAfterRendering() hook;
The reason behind that is that the focus method inside onAfterRendering could cause race conditions and may take focus away from the user in some edge cases.
We also decided to remove DayPicker.cy.tsx testing page, as ui5-daypicker is private component used for the composition of the ui5-calendar and it is NOT intended for single usage.