feat(ui5-menu): add placement api (#12554)
With this PR we introduce a new placement property to the ui5-menu.

This property allows developers to specify the menu's position relative to its trigger element, providing more control over the menu's placement in the UI.

The possible values for this property are Start, End, Top and Bottom, where Bottom is the default.

Example
To make the ui5-menu to open from the top add the following:

<ui5-menu placement="Top">
    ....
</ui5-menu>
image