fix(ui5-date-time-picker): correct initial focus on small screens (#12752)
Previously on initial open of the ui5-date-time-picker web component on smaller screen, the focus was being set on the ui5-segmented-button component. The reason behind this was that, the button was first in the DOM tree.

So now, instead of relying on automatic detection, we explicitly tell the ui5-responsive-popover which element should receive focus, by utilizing the initialFocus property (in phone mode only).